インフラ層のものレボ案件集約を参照と更新で別ファイルに分けた
