+++ SynchronizeMonorevoDeliveryDates.code-workspace
@@ -8,12 +8,14 @@
 		"cSpell.words": [
 			"divs",
 			"fropositions",
+			"goquery",
 			"gorm",
 			"jobbook",
 			"monorevo",
 			"orderdb",
 			"postable",
 			"prefs",
+			"Puerkito",
 			"sclevine",
 			"sqlserver",
 			"stretchr",
