+++ domain/monorevo/proposition_fetchposter.go
@@ -12,15 +12,18 @@ type Poster interface {
 	PostRange([]Proposition) error
 }
 
+// ものレボ案件のドメインモデル
 type Proposition struct {
 	WorkedNumber string
+	Det          string
 	DeliveryDate time.Time
 }
 
-func NewProposition(w string, d time.Time) *Proposition {
+func NewProposition(warkNumber string, det string, deliveryDate time.Time) *Proposition {
 	return &Proposition{
-		WorkedNumber: w,
-		DeliveryDate: d,
+		WorkedNumber: warkNumber,
+		Det:          det,
+		DeliveryDate: deliveryDate,
 	}
 }
 
@@ -28,7 +31,8 @@ func NewProposition(w string, d time.Time) *Proposition {
 // 参考: https://shiimanblog.com/engineering/functional-options-pattern/
 type Options struct {
 	WorkedNumber string
-	deliveryDate time.Time
+	Det          string
+	DeliveryDate time.Time
 }
 
 type Option func(*Options)
@@ -37,7 +41,8 @@ func TestPropositionCreate(options ...Option) *Proposition {
 	// デフォルト値設定
 	opts := &Options{
 		WorkedNumber: "99A-1234",
-		deliveryDate: time.Now(),
+		Det:          "1",
+		DeliveryDate: time.Now(),
 	}
-	return NewProposition(opts.WorkedNumber, opts.deliveryDate)
+	return NewProposition(opts.WorkedNumber, opts.Det, opts.DeliveryDate)
 }
